requirement.txt
selenium
Pillow
pytesseract
